+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPR833XZZS0BVDQNR93E09W0"
+        "01KPRXNT807729KS4JEX9C2D0N"
     )
     r"""The unique identifier of the data source"""