+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPR833XY4ZPYA9K7WEBE0ST3"
+        "01KPRXNT7ZDSAVXPSABZG3R54W"
     )
     r"""The unique identifier of the data source"""