+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPR833Y11N0SMWYE271E21DM"
+        "01KPRXNT8317EDKQCZZ4SME7AS"
     )
     r"""The unique identifier of the data source"""