+++ packages/orq-rc/RELEASES.md
@@ -7189,3 +7189,13 @@ Based on:
 - [python v4.8.0-rc.42] packages/orq-rc
 ### Releases
 - [PyPI v4.8.0-rc.42] https://pypi.org/project/orq-ai-sdk/4.8.0-rc.42 - packages/orq-rc
+
+## 2026-04-21 21:06:30
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.9 (2.881.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.8.0-rc.43] packages/orq-rc
+### Releases
+- [PyPI v4.8.0-rc.43] https://pypi.org/project/orq-ai-sdk/4.8.0-rc.43 - packages/orq-rc